% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attempts_parse_flash.R
\name{attempts_parse_flash}
\alias{attempts_parse_flash}
\title{Collects attempts within \code{tf_parse}}
\usage{
attempts_parse_flash(text)
}
\arguments{
\item{text}{output of \code{read_results} with row numbers appended by \code{add_row_numbers}}
}
\value{
returns a dataframe with split times and row numbers
}
\description{
Takes the output of \code{read_results} and, inside of \code{tf_parse}, extracts jump/throw attempts and associated row numbers
}
\seealso{
\code{attempts_parse_flash} runs inside \code{\link{flash_parse}} on the output of \code{\link{read_results}} with row numbers from \code{\link{add_row_numbers}}
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
