% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attempts_remove.R
\name{attempts_remove}
\alias{attempts_remove}
\title{Collects flight attempts within \code{tf_parse}}
\usage{
attempts_remove(df)
}
\arguments{
\item{df}{dataframe with jump attempt columns containing (X, O, PASS etc) and other columns}
}
\value{
returns a dataframe with the attempt columns removed
}
\description{
Takes the output of \code{read_results} and, inside of \code{tf_parse}, extracts vertical jump attempts and associated row numbers
}
\seealso{
\code{attempts_remove} runs inside \code{\link{flash_parse}}
}
