% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attempts_split_cols.R
\name{attempts_split_cols}
\alias{attempts_split_cols}
\title{Creates new columns for splitting attempts strings}
\usage{
attempts_split_cols(i, data, new_cols, old_cols)
}
\arguments{
\item{i}{iterative value}

\item{data}{output from \code{tf_parse}}

\item{new_cols}{a list of new column names to make}

\item{old_cols}{a list of old columns to split}
}
\value{
returns a data frame with Round_X_Attempts columns split into
individual attempts inside \code{tf_parse}
}
\description{
Given a data frame with columns "Round_1_Attempts" it will produce three
columns, for each of the attempts in round 1 (Round_1_Attempt_1,
Round_1_Attempt_2 etc.)
}
\seealso{
\code{attempts_split_cols} is a helper function inside
\code{attempts_split}
}
