% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_parse.R
\name{event_parse}
\alias{event_parse}
\title{Pulls out event labels from text}
\usage{
event_parse(text)
}
\arguments{
\item{text}{output from \code{read_results} followed by \code{add_row_numbers}}
}
\value{
returns a dataframe with event names and row numbers to eventually be recombined with track and field results inside \code{tf_parse}
}
\description{
Locates event labels in text of results output from \code{read_results} and their associated row numbers.  The resulting dataframe is joined back into results to include event names
}
\seealso{
\code{event_parse} is a helper function inside \code{\link{tf_parse}}
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
