% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_down.R
\name{fill_down}
\alias{fill_down}
\title{Fills NA values with previous non-NA value}
\usage{
fill_down(x)
}
\arguments{
\item{x}{a list having some number of non-NA values}
}
\value{
a list where NA values have been replaced with the closest previous non-NA value
}
\description{
This is a base approximation of \code{tidyr::fill()}
}
\seealso{
\code{fill_down} is a helper function inside \code{lines_sort}
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
