% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attempts_split_long.R
\name{flash_attempts_split_long_helper}
\alias{flash_attempts_split_long_helper}
\title{Creates new columns for splitting attempts strings in long format}
\usage{
flash_attempts_split_long_helper(data)
}
\arguments{
\item{data}{output from \code{tf_parse}}
}
\value{
returns a data frame with Round_X_Attempts columns split into
individual rows
}
\description{
Given a data frame with columns "Round_1_Attempts" it will produce three
rows, for each of the attempts in flight 1
}
\seealso{
\code{attempts_split_long_helper} is a helper function inside
\code{attempts_split_long}
}
