% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_clean_distance_events.R
\name{flash_clean_distance_events}
\alias{flash_clean_distance_events}
\alias{distance_events}
\title{Cleans distance events}
\usage{
flash_clean_distance_events(df, wide_format_distance = wide_format_clean)

distance_events(df, wide_format_distance = wide_format_clean)
}
\arguments{
\item{df}{a data frame of distance event data from Flash Results}

\item{wide_format_distance}{should df be presented in wide format (default is
\code{FALSE})?}
}
\value{
a cleaned version of df
}
\description{
Cleans distance event results pulled from Flash Results html tables.
Distance events are generally those with lengths of 400m or greater.  Can
present cleaned data in wide or long format.
}
\seealso{
\code{flash_clean_distance_events} is a helper function inside
\code{\link{flash_parse_table}}
}
