% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_clean_events.R
\name{flash_clean_events_helper}
\alias{flash_clean_events_helper}
\title{Applies appropriate event cleaning function}
\usage{
flash_clean_events_helper(
  df_helper = df,
  wide_format_clean_helper = wide_format_clean
)
}
\arguments{
\item{df_helper}{a data frame of vertical event data from Flash Results}

\item{wide_format_clean_helper}{should df be presented in wide format
(default is \code{FALSE})?}
}
\value{
a cleaned version of df
}
\description{
Used to apply appropriate cleaning function based on event name
}
\seealso{
\code{flash_clean_events_helper} is a helper function inside
\code{\link{flash_clean_events}}
}
