% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_clean_horizontal_events.R
\name{flash_clean_horizontal_events}
\alias{flash_clean_horizontal_events}
\alias{horizontal_events}
\title{Cleans horizontal events}
\usage{
flash_clean_horizontal_events(df, wide_format_horizontal = wide_format_clean)

horizontal_events(df, wide_format_horizontal = wide_format_clean)
}
\arguments{
\item{df}{a data frame of horizontal event data from Flash Results}

\item{wide_format_horizontal}{should df be presented in wide format (default
is \code{FALSE})?}
}
\value{
a cleaned version of df
}
\description{
Cleans horizontal event results pulled from Flash Results html tables.  Can
present cleaned data in wide or long format.
}
\seealso{
\code{flash_clean_horizontal_events} is a helper function inside
\code{\link{flash_parse_table}}
}
