% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_clean_relay_events.R
\name{flash_clean_relay_events}
\alias{flash_clean_relay_events}
\alias{relay_events}
\title{Cleans relay events}
\usage{
flash_clean_relay_events(df, wide_format_relay)

relay_events(df, wide_format_relay)
}
\arguments{
\item{df}{a data frame of relay event data from Flash Results}

\item{wide_format_relay}{should df be presented in wide format (default is
\code{FALSE})?}
}
\value{
a cleaned version of df
}
\description{
Cleans results pulled from Flash Results html tables for relay events. Can
present cleaned data in wide or long format.
}
\seealso{
\code{flash_clean_relay_events} is a helper function inside
\code{\link{flash_parse_table}}
}
