% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_clean_sprint_events.R
\name{flash_clean_sprint_events}
\alias{flash_clean_sprint_events}
\title{Cleans sprint events}
\usage{
flash_clean_sprint_events(df)
}
\arguments{
\item{df}{a data frame of sprint event data from Flash Results}

\item{wide_format_clean}{should df be presented in wide format (default is \code{FALSE})?}
}
\value{
a cleaned version of df
}
\description{
Cleans sprint event results pulled from Flash Results html tables.  Distance events are generally those with lengths of less than 400m.  Can present cleaned data in wide or long format.
}
\seealso{
\code{flash_clean_sprint_events} is a helper function inside \code{\link{flash_parse_table}
}
\author{
Gregory A. Pilgrim \email{gpilgrim2670@gmail.com} and George M. Perry
}
