% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_clean_vertical_events.R
\name{flash_clean_vertical_events}
\alias{flash_clean_vertical_events}
\alias{vertical_events}
\title{Cleans vertical events}
\usage{
flash_clean_vertical_events(df, wide_format_vertical = wide_format_clean)

vertical_events(df, wide_format_vertical = wide_format_clean)
}
\arguments{
\item{df}{a data frame of vertical event data from Flash Results}

\item{wide_format_vertical}{should df be presented in wide format (default is \code{FALSE})?}
}
\value{
a cleaned version of df
}
\description{
Cleans vertical event results pulled from Flash Results html tables.  Can present cleaned data in wide or long format.
}
\seealso{
\code{flash_clean_vertical_events} is a helper function inside \code{\link{flash_parse_table}}
}
\author{
Gregory A. Pilgrim \email{gpilgrim2670@gmail.com} and George M. Perry
}
