% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_col_names.R
\name{flash_col_names}
\alias{flash_col_names}
\title{Regularizes column names from Flash Results}
\usage{
flash_col_names(df)
}
\arguments{
\item{df}{a data frame or list of data frames containing event data from
Flash Results}
}
\value{
a version of df with split column names renamed
}
\description{
Split columns have many different naming conventions within Flash Results.
This function attempts to enforce one convention, "Split_XXX" where XXX are
digits representing distance in meters
}
