% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_col_names.R
\name{flash_col_names_helper}
\alias{flash_col_names_helper}
\title{Helper Function for regularizing column names from Flash Results}
\usage{
flash_col_names_helper(old_names)
}
\arguments{
\item{old_names}{a list of column names to be reformatted}
}
\value{
a list of strings containing corrected split column names
}
\description{
Helper Function for regularizing column names from Flash Results
}
