% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_correct_column_overshoot.R
\name{flash_correct_column_overshoot}
\alias{flash_correct_column_overshoot}
\title{Corrects column index overshoots when naming columns based on their contents}
\usage{
flash_correct_column_overshoot(x = NA, df)
}
\arguments{
\item{x}{a column position index}

\item{df}{a data frame with missing column names}
}
\value{
a correct index for column x
}
\description{
When naming columns based on the contents of a data frame the position of a
particular term, e.g. "Athlete" is used to name a column "Athlete".  If there
is a blank row at the top of the data frame then the position of "Athlete"
will be offset by the number of columns in the data frame.  This function
corrects for that.
}
\seealso{
\code{flash_correct_column_overshoot} is a helper function inside
\code{\link{flash_parse_table}}
}
