% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_date_parse.R
\name{flash_date_parse}
\alias{flash_date_parse}
\title{Pulls out date from text of flash results html page}
\usage{
flash_date_parse(text)
}
\arguments{
\item{text}{raw text of an event page from Flash Results}
}
\value{
a one element list containing the date of the event
}
\description{
Locates an date in text of results from a flash results html page for a given
event.
}
\seealso{
\code{flash_date_parse} is a helper function inside
\code{\link{flash_parse_table}}
}
