% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_event_links.R
\name{flash_event_links}
\alias{flash_event_links}
\alias{meet_links}
\title{Collects all event result links from a meet landing page on flashresults.com}
\usage{
flash_event_links(meet_home)

meet_links(meet_home)
}
\arguments{
\item{meet_home}{a link to a meet landing page on flashresults.com}
}
\value{
returns a list of links to individual events from a given meet
}
\description{
Used in scraping flashresults.com.  Collects event result links from a meet
landing page
}
\examples{
\donttest{flash_event_links("https://flashresults.com/2019_Meets/Outdoor/07-25_USATF_CIS/")}

}
\author{
Gregory A. Pilgrim \email{gpilgrim2670@gmail.com} and George M.
Perry
}
