% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_event_parse.R
\name{flash_event_parse}
\alias{flash_event_parse}
\title{Pulls out event label from text of flash results html page}
\usage{
flash_event_parse(text)
}
\arguments{
\item{text}{raw text of an event page from Flash Results}
}
\value{
a one element list containing the name of the event
}
\description{
Locates an event label in text of results from a flash results html page for a given event.
}
\seealso{
\code{flash_event_parse} is a helper function inside \code{\link{flash_parse_table}}
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
