% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_extract_details_links.R
\name{flash_extract_details_links}
\alias{flash_extract_details_links}
\alias{extract_details_links}
\title{Collects links to all detailed results links from a given event link on Flash
Results}
\usage{
flash_extract_details_links(link)

extract_details_links(link)
}
\arguments{
\item{link}{a link to an event landing page on flashresults.com}
}
\value{
returns list of links to corresponding detailed event result pages
}
\description{
Used in scraping flashresults.com.  Collects detailed results (often called
heat or flight results) from an associated event results landing page.
Detailed results often contain splits or attempts results.
}
\examples{
\donttest{flash_extract_details_links(
"https://flashresults.com/2015_Meets/Outdoor/06-25_USATF/008-3_compiled.htm")}

}
