% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_gender_parse.R
\name{flash_gender_parse}
\alias{flash_gender_parse}
\title{Pulls out gender label from text of flash results html page}
\usage{
flash_gender_parse(text)
}
\arguments{
\item{text}{raw text of an event page from Flash Results}
}
\value{
a one element list containing the gender of the event
}
\description{
Locates an gender label in text of results from a flash results html page for a given event.
}
\seealso{
\code{flash_gender_parse} is a helper function inside \code{\link{flash_parse_table}}
}
