% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_pivot_longer.R
\name{flash_pivot_longer}
\alias{flash_pivot_longer}
\title{Converts Flash Results from wide to long format}
\usage{
flash_pivot_longer(df, varying)
}
\arguments{
\item{df}{a data frame or list of data frames containing event data from
Flash Results}

\item{varying}{names of columns containing varying information (i.e.
splits)}
}
\value{
a version of df with split column values as \code{Split_Time} and split
column names as \code{Split_Distance}
}
\description{
Used to convert multiple split columns to two columns, \code{Split_Time} and
\code{Split_Distance}.  Effectively a T&F specific version of
\code{tidyr::pivot_longer} or \code{base::reshape}
}
