% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_rebuild_event_table.R
\name{flash_rebuild_event_table}
\alias{flash_rebuild_event_table}
\title{Rebuilds tables that \code{rvest::html_table} can't parse inside of \code{\link{flash_parse_table}}}
\usage{
flash_rebuild_event_table(event_url_rebuild)
}
\arguments{
\item{event_url_rebuild}{a link to an event page on flashresults.com}
}
\value{
returns a data frame of event results
}
\description{
Extracts individual \code{td} and \code{th} elements from html tables on Flash Results that cannot be parsed by code{rvest::html_table} (due to formatting issues in the html code)
}
\seealso{
\code{rebuild_event_table} is a helper function inside \code{\link{flash_parse_table}}
}
