% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_year_links.R
\name{flash_year_links}
\alias{flash_year_links}
\title{Collects all meet links from a given year on Flash Results}
\usage{
flash_year_links(flash_year)
}
\arguments{
\item{flash_year}{a link to a year landing page on flashresults.com}
}
\value{
returns a data frame with meet names, dates, locations, and links to flash results
}
\description{
Used in scraping flashresults.com.  Collects meet names, dates, and locations along with a link the the assoicated results landing page.
}
\examples{
\donttest{flash_year_links("https://flashresults.com/2015results.htm")}
}
\author{
George M. Perry
}
