% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attempts_split_long.R
\name{hytek_attempts_split_long_helper}
\alias{hytek_attempts_split_long_helper}
\title{Creates new columns for splitting attempts strings in long format}
\usage{
hytek_attempts_split_long_helper(i, data, old_cols)
}
\arguments{
\item{i}{output from \code{read_results} followed by \code{add_row_numbers}}

\item{data}{output from \code{tf_parse}}

\item{old_cols}{a list of old columns to split}
}
\value{
returns a data frame with Round_X_Attempts columns split into
individual rows
}
\description{
Given a data frame with columns "Round_1_Attempts" it will produce three
rows, for each of the attempts in flight 1
}
\seealso{
\code{attempts_split_long_helper} is a helper function inside
\code{attempts_split_long}
}
