% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytek_parse.R
\name{hytek_parse}
\alias{hytek_parse}
\title{Parses Hytek format track and field results inside \code{tf_parse}}
\usage{
hytek_parse(
  hytek_file = file,
  hytek_relay_athletes = relay_athletes,
  hytek_flights = flights,
  hytek_flight_attempts = flight_attempts,
  hytek_split_attempts = split_attempts
)
}
\arguments{
\item{hytek_file}{data with row numbers added}

\item{hytek_relay_athletes}{should \code{tf_parse} try to include the names of
relay athletes for relay events?  Names will be listed in new columns
"Relay-Athlete_1", "Relay_Athlete_2" etc.  Defaults to \code{FALSE}.}

\item{hytek_flights}{should \code{tf_parse} try to include flights for
jumping/throwing events?  Please note this will add a significant number of
columns to the resulting dataframe.  Defaults to \code{FALSE}.}

\item{hytek_flight_attempts}{should \code{tf_parse} try to include flights results
(i.e. "PASS", "X", "O") for high jump and pole value events?  Please note
this will add a significant number of columns to the resulting dataframe.
Defaults to \code{FALSE}}

\item{hytek_split_attempts}{should \code{tf_parse} split attempts from each flight
into separate columns?  For example "XXO" would result in three columns,
one for "X', another for the second "X" and third for "O".  There will be a
lot of columns.  Defaults to \code{FALSE}}
}
\value{
a data frame of track and field results
}
\description{
Outputs a data frame of track and field results
}
\seealso{
\code{hytek_parse} is runs inside
\code{\link{tf_parse}}
}
