% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_link_broken.R
\name{is_link_broken}
\alias{is_link_broken}
\title{Determines if a link is valid}
\usage{
is_link_broken(link_to_test, duration = 1)
}
\arguments{
\item{link_to_test}{a link}

\item{duration}{the lowest row number}
}
\value{
\code{TRUE} if the link works, \code{FALSE} if it fails
}
\description{
Used in testing links to external data, specifically inside of internal package tests.
Attempts to connect to link for the length of duration (in s).  If it fails it returns \code{FALSE}
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
