% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_transform.R
\name{list_transform}
\alias{list_transform}
\title{Transform list of lists into data frame}
\usage{
list_transform(x)
}
\arguments{
\item{x}{a list of lists, with all sub-lists having the same length}
}
\value{
a data frame where each sub-list is a row and each element of that sub-list is a column
}
\description{
Converts list of lists, with all sub-lists having the same number of elements into a data frame where each sub-list is a row and each element a column
}
\seealso{
\code{list_transform} is a helper function used inside of \code{tf_parse} and \code{event_parse}
}
