% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_format_helper.R
\name{math_format_helper}
\alias{math_format_helper}
\title{Helper function for formatting mm:ss.th times as seconds}
\usage{
math_format_helper(x)
}
\arguments{
\item{x}{A character vector of time(s) in track format (e.g. 1:35.93) to be converted to seconds (95.93)}
}
\value{
a numeric value representing a time or distance.  Units are not included
}
\description{
Helper function for formatting mm:ss.th times as seconds
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
