% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_conversion_helper.R
\name{metric_conversion_helper}
\alias{metric_conversion_helper}
\title{Converts distances in feet-inches to meters}
\usage{
metric_conversion_helper(x)
}
\arguments{
\item{x}{A character vector of distance(s) to be converted from feet-inches to meters}
}
\value{
a numeric value representing a number of meters.  Units are not included
}
\description{
Converts distances in feet-inches to meters
}
