% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_unneeded_attempts.R
\name{remove_unneeded_attempts}
\alias{remove_unneeded_attempts}
\title{Collects removes unneeded attempts columns within \code{tf_parse}}
\usage{
remove_unneeded_attempts(x)
}
\arguments{
\item{x}{dataframe with columns called both "Attempt_X" and "Attempt_X_Result" where X is a number}
}
\value{
returns a dataframe where Attempt_X columns that do no have a corresponding Attempt_X_Result have been removed
}
\description{
Inside of \code{tf_parse} & \code{tf_parse}, removes attempts columns that do not have an assoicated attempt_result
}
\seealso{
\code{remove_unneeded_attempts} runs inside \code{\link{flash_parse}} & \code{\link{tf_parse}}
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
