% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_unneeded_flights.R
\name{remove_unneeded_flights}
\alias{remove_unneeded_flights}
\title{Removes unneeded flights columns within \code{tf_parse}}
\usage{
remove_unneeded_flights(x)
}
\arguments{
\item{x}{dataframe with columns called both "Flight_X" and "Flight_X_Results"
where X is a number}
}
\value{
returns a dataframe where Flight_X columns that do not have a
corresponding Flight_X_Results have been removed
}
\description{
Inside of \code{tf_parse} & \code{tf_parse}, removes flight columns that do
not have an associated flight_attempts column
}
\seealso{
\code{remove_unneeded_flights} runs inside \code{\link{flash_parse}}
& \code{\link{tf_parse}}
}
