% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits_parse.R
\name{splits_parse}
\alias{splits_parse}
\title{Collects splits within \code{tf_parse}}
\usage{
splits_parse(text, split_len = 1)
}
\arguments{
\item{text}{output of \code{read_results} with row numbers appended by
\code{add_row_numbers}}

\item{split_len}{the distance at which splits are measured}
}
\value{
returns a data frame with split times and row numbers
}
\description{
Takes the output of \code{read_results} and, inside of \code{tf_parse},
extracts split times and associated row numbers
}
\seealso{
\code{splits_parse} runs inside \code{\link{tf_parse}} on the
output of \code{\link{read_results}} with row numbers from
\code{\link{add_row_numbers}}
}
