% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_conversion.R
\name{standard_conversion}
\alias{standard_conversion}
\title{Formatting meters lengths as feet-inches}
\usage{
standard_conversion(x)
}
\arguments{
\item{x}{A character vector of distance(s) in meters format ("3.85m") , to be converted to meters ("12-07.45")}
}
\value{
returns the value of the string \code{x} which represents a distance in feet-inches
}
\description{
Takes a character string (or list) representing a length in meters format (e.g. "3.85m") and converts it to a distance in feet-inches ("12-07.45")
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
