% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_wind_from_rounds.R
\name{wind_from_rounds}
\alias{wind_from_rounds}
\title{Pulls Wind Data by Round from Horizontal Flash Table Results}
\usage{
wind_from_rounds(df)
}
\arguments{
\item{df}{a data frame containing results with wind data included in round
columns.}
}
\value{
a data frame with all wind data in separate (tidy) columns
}
\description{
In some Flash Table results for horizontal events (long jump, triple jump,
throwing events), a wind value is listed for each round/attempt.  This
function pulls out those wind values into columns called "Round_1_Wind" (if
the round data is in a column called Round_1)
}
\seealso{
\code{wind_from_rounds} is a helper function inside
\code{\link{flash_clean_horizontal_events}}
}
