% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_wind_from_rounds.R
\name{wind_from_rounds_helper}
\alias{wind_from_rounds_helper}
\title{Helper function for extracting wind data from round columns}
\usage{
wind_from_rounds_helper(df = df, i, round_cols, ...)
}
\arguments{
\item{df}{a data frame containing round columns with both results and wind
data}

\item{i}{list of values to iterate along}

\item{round_cols}{list of columns containing results and wind values by round}

\item{...}{other arguments as needed}
}
\value{
a list of data frames with all wind data for each round in a separate
(tidy) column
}
\description{
Helper function for extracting wind data from round columns
}
