% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytek_wind_parse.R
\name{wind_parse_hytek}
\alias{wind_parse_hytek}
\title{Collects splits within \code{tf_parse}}
\usage{
wind_parse_hytek(text)
}
\arguments{
\item{text}{output of \code{read_results} with row numbers appended by
\code{add_row_numbers}}
}
\value{
returns a data frame with wind speeds and row numbers
}
\description{
Takes the output of \code{read_results} and, inside of \code{tf_parse},
extracts split times and associated row numbers
}
\seealso{
\code{wind_parse_hytek} runs inside \code{\link{hytek_parse}} on the
output of \code{\link{read_results}} with row numbers from
\code{\link{add_row_numbers}}
}
