% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_row_numbers.R
\name{add_row_numbers}
\alias{add_row_numbers}
\title{Add row numbers to raw results}
\usage{
add_row_numbers(text)
}
\arguments{
\item{text}{output from \code{read_results}}
}
\value{
returns a data frame with event names and row numbers to eventually be
recombined with T&F results inside \code{swim_parse}
}
\description{
Takes the output of \code{read_results} and adds row numbers to it
}
\seealso{
\code{add_row_numbers} is a helper function inside
\code{\link{tf_parse}}
}
