% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attempts_split.R
\name{attempts_split}
\alias{attempts_split}
\title{Creates new columns of split attempts strings}
\usage{
attempts_split(data_to_split)
}
\arguments{
\item{data_to_split}{output from \code{read_results} followed by
\code{add_row_numbers}}
}
\value{
returns a data frame with Flight_X_Attempts columns split into
individual attempts inside \code{tf_parse}
}
\description{
Given a data frame with columns "Flight_1_Attempts" it will output three
columns, for each of the attempts in flight 1 (Flight_1_Attempt_1,
Flight_1_Attempt_2 etc.)
}
\seealso{
\code{attempts_split} is a helper function inside
\code{\link{tf_parse}}
}
