% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attempts_split_long.R
\name{attempts_split_long}
\alias{attempts_split_long}
\title{Creates new rows of split attempts strings (long format change)}
\usage{
attempts_split_long(data_to_split)
}
\arguments{
\item{data_to_split}{output from \code{read_results} followed by
\code{add_row_numbers}}
}
\value{
returns a data frame with Round_X_Attempts columns split into
individual attempts as rows
}
\description{
Given a data frame with columns "Round_1_Attempts" it will create three new
rows, one for each of the attempts in round 1
}
\examples{
\donttest{
df <- tf_parse( read_results(
"https://www.flashresults.com/2018_Meets/Outdoor/04-20_DukeInvite/014-1.pdf"
), rounds = TRUE, round_attempts = TRUE, )

 df \%>\% attempts_split_long()
}
}
