% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_left.R
\name{fill_left}
\alias{fill_left}
\title{Shifts non-NA values to left in data frame}
\usage{
fill_left(df)
}
\arguments{
\item{df}{a data frame having some NA values}
}
\value{
a data frame where all values have been pushed left, replacing NAs,
and all columns containing only \code{NA} values have been removed
}
\description{
Moves non-NA data left into NA spaces, then removes all columns that contain
only NA values
}
\seealso{
\code{fill_left} is a helper function inside \code{lines_sort}
}
