% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_rebuild_event_table.R
\name{flash_rebuild_event_table}
\alias{flash_rebuild_event_table}
\title{Rebuilds tables that html_table can't parse within flash_parse_table}
\usage{
flash_rebuild_event_table(event_url_rebuild)
}
\arguments{
\item{event_url_rebuild}{a link to an event page on flashresults.com}
}
\value{
returns a dataframe of event results
}
\description{
Extracts individual td and th elements from html tables on Flash Results that cannot be parsed by html_parse (due to formatting issues in the html code)
}
\seealso{
\code{rebuild_event_table} is a helper function inside \code{\link{flash_parse_table}}
}
\author{
Gregory A. Pilgrim
}
