% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_conversion_helper.R
\name{standard_conversion_helper}
\alias{standard_conversion_helper}
\title{Converts distances in meters to feet-inches}
\usage{
standard_conversion_helper(x)
}
\arguments{
\item{x}{A character vector of distance(s) to be converted from meters to
feet-inches}
}
\value{
a character vector in feet-inches
}
\description{
Converts distances in meters to feet-inches
}
