% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_parse.R
\name{tf_parse}
\alias{tf_parse}
\title{Reads track and field results into a list of strings in preparation for parsing with \code{tf_parse}}
\usage{
tf_parse(
  file,
  avoid = avoid_default,
  typo = typo_default,
  replacement = replacement_default
)
}
\arguments{
\item{file}{a .pdf or .html file (could be a url) where containing track and field results.  Must be formatted in a "normal" fashion - see vignette}

\item{avoid}{xxx}

\item{typo}{xxx}

\item{replacement}{xxx}
}
\value{
a dataframe
}
\description{
Outputs list of strings to be processed by \code{tf_parse}
}
\seealso{
\code{tf_parse} is meant to be preceeded by \code{\link{read_results}}
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
